// src/modules/schema/types.ts
export interface Tests {
  passedAll: boolean
  passed: number
  failed: number
  totalTests: number
  successes: SuccessTest[]
  errors: ErrorTest[]
  time: string
}

export interface ErrorTest {
  class?: string
  method?: string
  type: 'invalid param' | 'invalid value' | 'missing value' | 'missing key'
  name: string
  expect: string
  received: any
  index?: number
  message: string
}

export interface SuccessTest {
  class?: string
  method?: string
  name: string
  expect: string
  index?: number
  received: any
}